Change when assembly stock availability is validated
Only validate the the total stock of all varaints in an order when
entering the checkout process. This was causing payment and shipment
states to be set to nil during `order.finalize!`. At that point the
stock has been allocated to the shipment but the order was still
validating if there was enough count on hand to fufill it again.